requirements.txt
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Faker==13.4.0
Flask==2.1.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
pdfkit==1.0.0
Pillow==9.1.0
platformdirs==2.5.2
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
